restructured question id files into seperate directories for maintainability
